teb: Use signed type for integers

+++ test/teb.c
@@ -22,19 +22,19 @@ main(void)
 
 #if __MCF_64_32(64, 32) != 32
     /* 64-bit */
-    assert(__MCF_teb_load_32(0x40) == GetCurrentProcessId());
-    assert(__MCF_teb_load_32(0x48) == GetCurrentThreadId());
+    assert(__MCF_teb_load_32(0x40) == (int32_t) GetCurrentProcessId());
+    assert(__MCF_teb_load_32(0x48) == (int32_t) GetCurrentThreadId());
 
     assert(TlsSetValue(tls_index, (void*) 0x0123456789ABCDEF));
     assert(__MCF_teb_load_ptr(0x1480 + tls_index * 8) == (void*) 0x0123456789ABCDEF);
-    assert(__MCF_teb_load_32(0x1480 + tls_index * 8) == 0x89ABCDEF);
-    assert(__MCF_teb_load_32(0x1480 + tls_index * 8 + 4) == 0x01234567);
+    assert(__MCF_teb_load_32(0x1480 + tls_index * 8) == (int32_t) 0x89ABCDEF);
+    assert(__MCF_teb_load_32(0x1480 + tls_index * 8 + 4) == (int32_t) 0x01234567);
 
     __MCF_teb_store_ptr(0x1480 + tls_index * 8, (void*) 0x13577531AFFADCCD);
     assert(TlsGetValue(tls_index) == (void*) 0x13577531AFFADCCD);
 
-    __MCF_teb_store_32(0x1480 + tls_index * 8, 0x21436587);
-    __MCF_teb_store_32(0x1480 + tls_index * 8 + 4, 0xA9CBED0F);
+    __MCF_teb_store_32(0x1480 + tls_index * 8, (int32_t) 0x21436587);
+    __MCF_teb_store_32(0x1480 + tls_index * 8 + 4, (int32_t) 0xA9CBED0F);
     assert(TlsGetValue(tls_index) == (void*) 0xA9CBED0F21436587);
 
     void* rtl_user_process_parameters = *(void**) ((char*) __MCF_peb() + 0x20);
@@ -46,17 +46,17 @@ main(void)
 
 #if __MCF_64_32(64, 32) != 64
     /* 32-bit */
-    assert(__MCF_teb_load_32(0x20) == GetCurrentProcessId());
-    assert(__MCF_teb_load_32(0x24) == GetCurrentThreadId());
+    assert(__MCF_teb_load_32(0x20) == (int32_t) GetCurrentProcessId());
+    assert(__MCF_teb_load_32(0x24) == (int32_t) GetCurrentThreadId());
 
     assert(TlsSetValue(tls_index, (void*) 0x89ABCDEF));
     assert(__MCF_teb_load_ptr(0x0E10 + tls_index * 4) == (void*) 0x89ABCDEF);
-    assert(__MCF_teb_load_32(0x0E10 + tls_index * 4) == 0x89ABCDEF);
+    assert(__MCF_teb_load_32(0x0E10 + tls_index * 4) == (int32_t) 0x89ABCDEF);
 
     __MCF_teb_store_ptr(0x0E10 + tls_index * 4, (void*) 0x13577531);
     assert(TlsGetValue(tls_index) == (void*) 0x13577531);
 
-    __MCF_teb_store_32(0x0E10 + tls_index * 4, 0x21436587);
+    __MCF_teb_store_32(0x0E10 + tls_index * 4, (int32_t) 0x21436587);
     assert(TlsGetValue(tls_index) == (void*) 0x21436587);
 
     void* rtl_user_process_parameters = *(void**) ((char*) __MCF_peb() + 0x10);
